vote-interface: make `deserialize_into` generic
